services/artifact_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
python-multipart>=0.0.9
sentence-transformers>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
striprtf>=0.0.26
pytesseract>=0.3.10
Pillow>=10.0.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.40.0
yt-dlp>=2024.1.0
ollama>=0.4.0
anthropic>=0.40.0
openai>=1.50.0
